Finally fix that one compiler warning
bench: 5746232

+++ src/external/fathom/tbprobe.cpp
@@ -2460,7 +2460,7 @@ void tb_expand_mate(Pos *pos, struct TbRootMove *move, Value moveScore, unsigned
   }
 
   // Now try to expand until the actual mate.
-  if (popcount(pos->white | pos->black) <= cardinalityDTM) {
+  if (static_cast<unsigned int>(popcount(pos->white | pos->black)) <= cardinalityDTM) {
     while (v != -TB_VALUE_MATE && move->pvSize < TB_MAX_PLY) {
       v = v > 0 ? -v - 1 : -v + 1;
       wdl = -wdl;
